pyproject.toml
[project]
name = "eowautomater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.43",
    "pydantic-ai>=0.0.12",
]
